refactor(rpc): move helpers to individual submodules
