+++ crates/rpc/src/server/api/get_account.rs
@@ -1,10 +1,12 @@
 use miden_node_proto::domain::account::{AccountRequest, AccountStorageRequest, SlotData};
 use miden_node_proto::generated as proto;
+use miden_node_store::GetAccountError;
 use miden_node_utils::limiter::QueryParamStorageMapKeyTotalLimit;
 use miden_node_utils::tracing::OpenTelemetrySpanExt;
+use tonic::Status;
 use tracing::{Span, debug, info_span};
 
-use super::{COMPONENT, RpcService, check, get_account_error_to_status};
+use super::{COMPONENT, RpcService, check};
 
 #[tonic::async_trait]
 impl proto::server::rpc_api::GetAccount for RpcService {
@@ -49,5 +51,20 @@ impl proto::server::rpc_api::GetAccount for RpcService {
         let account_data =
             self.store.get_account(request).await.map_err(get_account_error_to_status)?;
         Ok(account_data.into())
+    }
+}
+
+// HELPERS
+// ================================================================================================
+
+fn get_account_error_to_status(err: GetAccountError) -> Status {
+    let message = err.to_string();
+    match err {
+        GetAccountError::DatabaseError(err) => super::database_error_to_status(&err),
+        GetAccountError::DeserializationFailed(_)
+        | GetAccountError::AccountNotFound(..)
+        | GetAccountError::AccountNotPublic(_)
+        | GetAccountError::UnknownBlock(_)
+        | GetAccountError::BlockPruned(_) => Status::invalid_argument(message),
     }
 }
