+++ crates/rpc/src/server/api/status.rs
@@ -1,15 +1,8 @@
+use miden_node_block_producer::{BlockProducerStatus, MempoolStats};
 use miden_node_proto::generated as proto;
 use tracing::debug;
 
-use super::{
-    COMPONENT,
-    Finality,
-    ProtoMempoolStats,
-    Request,
-    RpcMode,
-    RpcService,
-    block_producer_status_to_proto,
-};
+use super::{COMPONENT, Finality, ProtoMempoolStats, Request, RpcMode, RpcService};
 
 #[tonic::async_trait]
 impl proto::server::rpc_api::Status for RpcService {
@@ -51,5 +44,25 @@ impl proto::server::rpc_api::Status for RpcService {
             })),
             genesis_commitment: self.genesis_commitment.map(Into::into),
         })
+    }
+}
+
+// HELPERS
+// ================================================================================================
+
+fn block_producer_status_to_proto(status: BlockProducerStatus) -> proto::rpc::BlockProducerStatus {
+    proto::rpc::BlockProducerStatus {
+        version: status.version,
+        status: status.status,
+        chain_tip: status.chain_tip.as_u32(),
+        mempool_stats: Some(block_producer_mempool_stats_to_proto(status.mempool_stats)),
+    }
+}
+
+fn block_producer_mempool_stats_to_proto(stats: MempoolStats) -> proto::rpc::MempoolStats {
+    proto::rpc::MempoolStats {
+        unbatched_transactions: stats.unbatched_transactions,
+        proposed_batches: stats.proposed_batches,
+        proven_batches: stats.proven_batches,
     }
 }
